// apps/backend/src/activity-logs/enums/action-type.enum.ts
export enum ActionType {
  ADD_COMMENT = 'ADD_COMMENT',
  ASSIGN_ASSIGNEE = 'ASSIGN_ASSIGNEE',

  CHANGE_STATUS = 'CHANGE_STATUS',
  SET_DUE_DATE = 'SET_DUE_DATE',
  CHANGE_DUE_DATE = 'CHANGE_DUE_DATE',
  CHANGE_PRIORITY = 'CHANGE_PRIORITY',
  CHANGE_EXPECTED_RESULT = 'CHANGE_EXPECTED_RESULT',
  CHANGE_TITLE = 'CHANGE_TITLE',
  CHANGE_DESCRIPTION = 'CHANGE_DESCRIPTION',
  COMPLETION_REPORT_RESOLUTION = 'COMPLETION_REPORT_RESOLUTION',
}